requirements.txt
tweepy==3.6.0
python-decouple==3.1
ipython==6.5.0
django-rq==1.2.0
Django==2.1
Unipath==1.1
dj-database-url==0.5.0
django-extensions==2.0.7
django-suit==0.2.26
python-decouple==3.1
gunicorn==19.9.0
django-heroku==0.3.1
whitenoise==3.3.1
psycopg2-binary==2.7.5
raven==6.9.0
tqdm==4.24.0
